Extract shared adapters into a dedicated package
